refactor: move parts of arithmetic dataset generation to separate files
This commit moves the 'ExpressionGenerator', 'ASTNode, and 'ExpressionParser' classes to individual files. These changes make the code a little more structured.
